Give logger its own crate
